日常更新2024-4-3
1、添加了参数保存到板载FLASH的功能
2、基础的波形绘制
3、优化了串口的数据收发逻辑

+++ fyFLy/APPLICATION/inc/motoCtr.h
@@ -7,7 +7,7 @@
 
 #define PWM_MAX_RANGE 2000
 
-extern uint16_t outPwmValue[4];
+extern int32_t outPwmValue[4];
 
 void motoInit(void);
 void electricalCalibration(const int16_t *channel,_flyStateStruct *state);
